Simplify remote session module/handlers with tests (#28)
This change breaks apart the remote session handler into three separate handlers each with a specific function. It also moves any authentication and decisions as to how to handle the request to the module itself. This way, by the time a request has reached the handler, it knows it should just do its thing. Tests have also been included for these as well.