Call HttpResponse.FlushAsync instead of OutputStream.FlushAsync (#49)
A recent change changed things to call the flush on the stream rather than the request. This breaks the remote app session workflow as the output would not be flushed and so the core app would not be notified of the payload. Tests have been updated to verify this is called.