Switch serialization of session state to use BinaryWriter/Reader
This breaks apart the two parts of serializing session state with different needs:

- HttpSessionState/ISessionState itself: This is probably not as interesting to implement for people as it is just some data to understand the session state. The abstraction ISessionSerializer is still used for this, but the key serialization has been removed
- Keys and entries: This is the thing people are going to want to be able to customize. This change introduces an abstraction focused on this: ISessionKeySerializer

With this change, someone who wants to really customize how session entries will be serialized can do whatever they want. They could chain together ISessionKeySerializer instances, or pick some strategy that makes sense for how they generate keys (i.e. dynamically on a pattern). Without this change, in order to do that, they would also have to implement the transformation for the whole HttpSessionState which is much more complicated.

Another nice thing about this structure, is that we can move majority of the session infrastructure into the core system adapters. An additional level would still be needed for System.Text.Json usage, but it could be a general package of all the JSON formatting people may want to opt into.

The final thing this change does is change the ISessionState a bit so that it can be available on .NET Framework as well without new dependencies. This is done by replacing IAsyncDisposable and ValueTask with IDisposable and Task.