Add HttpResponse APIs (ClearHeaders/WriteFile/TransmitFile/etc) (#102)
This change adds a few APIs that were missing from HttpResponse.

SubStatusCode is a placeholder for the value, but does not actually publish it anywhere.

TransmitFile/WriteFile: These APIs are essentially the same, except that they had different memory characteristics. TransmitFile did not buffer, while WriteFile did. On ASP.NET Core, these will just delegate to IHttpResponseBodyFeature.SendFileAsync.

NOTE: This implementation will call Task.GetAwaiter() which should be avoided. We could potentially expose an async API for this on the ASP.NET Core side as a stepping stone if we want.

Part of #91