By default exclude all code from .NET Standard
The .NET Standard build only needs to have the references file compiled in. Everything else is infrastructure and is only needed on .NET Core/Framework/Both.

This also moves the Authentication files under Adapters to follow the convention where top level is only System.Web types, and any supporting adapters are in the Adapters folder.