Consolidate session APIs into main library
This moves the session infrastructure into the main library. The current SessionState library existed because there was a dependency on System.Text.Json so we didn't want to force people to use that. However, now that can be opted into for a specific implementation of ISessionKeySerializer. This has been kept in the separate project which has been renamed to Microsoft.AspNetCore.SystemWebAdapters.SystemTextJson