Use a shared client for remote app connections (#178)
This change brings the configuration of the HttpClient into a shared
location for all the remote app connections. Now, any system that needs
to connect to the remote app can do so by just retrieving the client via
the IHttpClientFactory by name and the path/URI/backchannel will be
configured for them.