Add HttpRequest.FilePath and HttpRequest.PathInfo (#285)
This change adds the APIs for HttpRequest.FilePath|PathInfo and plumbs it through a new feature: IPathInfoFeature. The actual implementation of this is something tied to IHttpHandler, but this will default to the case in which no handler is registered. If we want to implement the behavior if a handler is present, the feature just needs to be added for the API to work as expected.