Refactor adapter features to be always available
Currently, the adapters replace a few features if endpoint metadata is configured for this to happen. However, this was intended to gate "expensive" operations, such as prebuffering the stream to enable InputStream. However, this now prevents APIs that are not "expensive" to be gated as well but are implemented in the feature.

This change will still keep the gating in place for the "expensive" operations, but enable APIs that are provided by the features to be available if possible without the endpoint metadata.