Keep IPrincipal user instances if not ClaimsPrincipal
There are instances where people are using non-ClaimsPrincipal users in ASP.NET Core. This change will allow us to track that IPrincipal and surface it as needed as a ClaimsPrincipal. This is mostly for use while migrating, as ideally they will switch to a ClaimsPrincipal based User, but will prevent information contained in the custom type from being lost.