Add build helpers for NuGet packages
This adds a target to the adapters that will allow making use of Content from older style NuGet packages that don't work with PackageReference. This is done by adding an attribute `CopyContent='true'` to the PackageReference node in the project file, which will cause all the content in the package to be passed into the Content ItemGroup and copied to the output directory.