Forward IHtmlString on .NET 8 (#357)
This adds a .NET 8 build target to the adapters so we can forward our implementation of IHtmlString to the new in-box version. With this, someone can use IHtmlString in .NET Standard, but will be forwarded on platforms that support it out of the box (i.e. .NET Framework and .NET 8+).

This change also reacts to some nullability changes that occured in .NET 8