Remove broken links (#364)
The README already links to the official docs. The pages these links point to do not exist as-is anymore and have been merged into the larger doc. So, we can just remove the links to these parts and the existing links will take people to the docs in general.

Fixes #363