Use an options type for hosting and runtime environment
This refactors the runtime environment to be configured via a user-available options type. By default, if running on IIS, this will be populated, but can easily be configured if needed.

This also adds HostingEnvironment properties that may be expected via the same options.

As part of this, a hosted service is added that will set the options as a static instance so it will be available for the lifetime of the service (since the APIs on HttpRuntime and HostingEnvironment were static). This hosted service will take care of clearing this value once the host is stopped.