Move HttpApplication middleware into UseSystemWebAdapters
This was initially moved into the startup filter to be added early on in the request pipeline because it was thought it was needed there for output caching. However, that does not seem to be the case (test has been added for that). So, now it can be moved into the standard `UseSystemWebAdapters()` so that it will be only added on requests that needed it (i.e. if the pipeline has been branched and `UseSystemWebAdapters()` is not added).