Use TimeProvider to set HttpContext.Timestamp
In order to use this, we need to move the timestamp to a feature so we can set this in the middleware where we have TimeProvider. As a benefit of this, we won't prematurely create a `System.Web.HttpContext` to initialize the time; instead, the feature will be created, but that's it.

Another change this requires is to add a .NET 8 build for the CoreServices library, which raises a few new warnings about some API changes that are addressed in this PR as well.