Separate HttpApplication feature insertion from invocation
This allows the features for HttpApplication to be insterted in the HttpContext at the beginning of the request, but will only be invoked in the `.UseSystemWebAdapters()` middleware call. This enables a request to potentially modify behavior of the features if needed before invoking the middleware.

A nice side effect of this is that an HttpApplication will only be created if it is actually used within a request, otherwise, it will never be retrieved (which is one of the more expensive parts of the emulated IIS pipeline as we have to assign an HttpApplication to a request).