Add System.Web.Security.MachineKey
This adds the API backed by the ASP.NET Core data protection APIs. In order to share code between ASP.NET Core and framework with the same behavior, the ASP.NET Core data protection APIs must be used on the framework side as well. A sample has been added that shows how to enable that (this is already a supported scenario for migration - this PR just hooks it up so the previous API can be used if needed).