Remove or fix APIs that were ported with different signatures
This is to fix a few APIs where the APIs ported did not match .NET Framework, so using those methods would throw at runtime. The only API removed in this is HttpRequest.LogonUserIdentity - we can investigate adding it back if needed. The rest of the APIs now match framework and tests are in place to ensure going forward no regressions or repeats of this issue occur.

Fixes #486