Remove public constructor from HttpSessionState (#553)
There was no public constructor for this on framework. This was added initially as a way to override the current session implementation. However, now it can be changed by setting `HttpContext.Features.Get<ISessionStateFeature>().State` and it will be propagated in the session.

This is a breaking change, but we're already doing that for a few other APIs for v2. This will help guide people to the recommended approach to override behavior.