Copy the keys from framework session state (#559)
It appears that some session providers on the framework side may not be able to iterate over the keys while grabbing items as it may throw an InvalidOperationException with the error: Collection was modified after the enumerator was instantiated. Even though it appears that this should not happen, we'll copy the keys over before grabbing items from the collection so this can't happen.