Use UsingMSBuildSDKSystemWeb property to avoid ASPIRE004 warning.
Rather than supressing the warning, set the property `_IsExecutable` on projects based on `MSBuild.SDK.SystemWeb` when calling the `GetTargetFrameworksWithPlatformForSingleTargetFramework` property which is used in the call to get `AdditionalPropertiesFromProject` which is used in the `GetNonExecutableReferences` task which returns project references which have this set to false.
`_ValidateAspireHostProjectResources` then triggers the `ASPIRE004` warning for each such project.