Update to Aspire 13 (#632)
This change required updating to .NET 10 for a number of components and as part of that it became clear that the CPM setup of separate package versions for samples/tests/product were very difficult to keep in sync. This consolidates it all into a single Directory.Package.props and updates the projects to use the latest aspire version.