Fix issue when enabling static user accessors (#639)
The RequestUserFeature implements the ASP.NET Core IHttpAuthenticationFeature as well as the adapter's IRequestUserFeature. We were only copying the user to the static accessors when the ASP.NET Core User was set not when the IRequestUserFeature one was. Since there's a single User property that stores it for both, this change moves the call to sync to the static accessors there.