Prevent issues around default scheme (#644)
There are a number of reports of various issues that appear to be because as of .NET 7, an authentication scheme is automatically the default scheme if there is only a single one. This makes it so that at times (such as YARP fallback), the authentication is automatically being run when there is no need.

A change between v1.4 and v2.0 caused a behavior change that manifests here as well.

This change does the following:

- Adds an empty authentication handler so that the the default scheme is never set by the new behavior. This handler is registered with a randomly generated GUID as it is not expected to be used at all
- Adds a check when the remote authentication is being used to check for more multiple Location headers (a common manifestation of this issue) with some guidance on what to do.